+++ docs/src/dev-docs/changelog.rst
@@ -24,7 +24,37 @@ changelog <https://keepachangelog.com/en/1.1.0/>`_ format. This project follows
 Unreleased
 ----------
 
-Version 2025.3 - 2025-03-11
+Fixed
+#####
+
+- ``NativePET`` can now evaluate on single-atom structures without crashing
+
+Version 2025.5 - 2025-04-13
+---------------------------
+
+Fixed
+#####
+
+- Fix more composition model issues
+
+Added
+#####
+
+- Update to ``metatensor-torch-0.7.5`` to allow training on ``non_conservative_forces``
+  and  ``non_conservative_stress`` targets
+- Add ``NativePET`` as a readable, efficient, backward-compatible PET implementation
+- Added Wandb logger
+- Save loss history in a ``.csv`` file
+
+Version 2025.4 - 2025-03-29
+---------------------------
+
+Changed
+#######
+
+- upgraded to ``metatensor.torch`` 0.7.4, which gives access to batched ASE evaluation
+
+Version 2025.3 - 2025-03-25
 ---------------------------
 
 Fixed